Refactor all components
Refactor components, so the library structure is more directory-oriented. Export all components, a plugin that install them (see `install` function), and an `iui` function that configures the lib but does not install components
