+++ package.json
@@ -7,7 +7,8 @@
     "serve": "vue-cli-service serve",
     "lib": "vue-cli-service build --target lib --name iui-vue ./src/index.ts",
     "build": "vue-cli-service build --dest ./docs",
-    "lint": "vue-cli-service lint --fix"
+    "lint": "vue-cli-service lint --fix",
+    "prepublish": "npm run lib"
   },
   "repository": {
     "type": "git",
