+++ src/index.ts
@@ -1,24 +1,16 @@
-export { default as Button } from './lib/components/Button.vue';
-
-export { default as Checkbox } from './lib/components/Checkbox.vue';
-export { default as CheckboxGroup } from './lib/components/CheckboxGroup.vue';
-
-export { default as Radio } from './lib/components/Radio.vue';
-export { default as RadioGroup } from './lib/components/RadioGroup.vue';
-
-export { default as Dropdown } from './lib/components/Dropdown.vue';
-export { default as DropdownSelect } from './lib/components/DropdownSelect.vue';
-
-export { default as Modal } from './lib/components/Modal.vue';
-
-export { default as Tabs } from './lib/components/Tabs.vue';
-export { default as TabPanel } from './lib/components/TabPanel.vue';
-
-export { default as Toggle } from './lib/components/Toggle.vue';
-
-export { default as CustomTransition } from './lib/components/Transition.vue';
-export { default as CustomTransitionGroup } from './lib/components/TransitionGroup.vue';
-
-export { default as Dropzone } from './lib/components/Dropzone.vue';
-
-export { default as iui } from './lib/utils/plugin';
+export {default as Button} from './lib/components/Button';
+export {default as Checkbox} from './lib/components/Checkbox';
+export {default as CheckboxGroup} from './lib/components/CheckboxGroup';
+export {default as Radio} from './lib/components/Radio';
+export {default as RadioGroup} from './lib/components/RadioGroup';
+export {default as Dropdown} from './lib/components/Dropdown';
+export {default as DropdownSelect} from './lib/components/DropdownSelect';
+export {default as Modal} from './lib/components/Modal';
+export {default as Tabs} from './lib/components/Tabs';
+export {default as TabPanel} from './lib/components/Tabs/panel';
+export {default as Toggle} from './lib/components/Toggle';
+export {default as Transition} from './lib/components/Transition';
+export {default as TransitionGroup} from './lib/components/TransitionGroup';
+export {default as Dropzone} from './lib/components/Dropzone';
+export {default as iui} from './lib/utils/plugin';
+export {default as install} from './lib/utils/install';
