+++ src/lib/components/Checkbox/index.vue
