+++ src/lib/components/Tabs/index.vue
